ref(core)!: Move `getComponentName` to browser-utils
This is only used there, and only really internally.
